feat: add Readme.Verify.Webhook() and update example server
